New apis

+++ code/avh.nufm.api/src/main/java/avh/nufm/api/model/APIInvoice.java
@@ -1,6 +1,8 @@
 package avh.nufm.api.model;
 
 
+import java.math.BigDecimal;
+
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
@@ -11,11 +13,11 @@ import lombok.Setter;
 public class APIInvoice {
 private String invoiceTo;
 private String deliverTo;
-private double quantity;
-private double price;
+private BigDecimal quantity;
+private BigDecimal price;
 private String description;
-private double subTotal;
-private double salesTax;
-private double invoiceTotal;
+private BigDecimal subTotal;
+private BigDecimal salesTax;
+private BigDecimal invoiceTotal;
 private String status;
 }
